Version 0.21 - Updated Spike.prefab & morem
The Spike now has a more accurate hit box, instead of the hit box being a square it is now a actual triangle!

There is also a score now! The score goes up every .3 seconds. When the game restarts, so does the score. The high score currently does not work because I ran out of time.